Essai de compilation et lancement sous windows (1ere partie)

+++ probleme0xx/probleme048.cpp
@@ -7,7 +7,7 @@ ENREGISTRER_PROBLEME(48, "Self powers") {
     //
     // Find the last ten digits of the series, 1^1 + 2^2 + 3^3 + ... + 1000^1000.
     mp_nombre resultat = 0;
-    for (size_t n = 1; n <= 1000; ++n) {
+    for (unsigned long n = 1; n <= 1000; ++n) {
         resultat += mp_nombre::puissance_modulaire(n, n, 10000000000LL);
         resultat %= 10000000000LL;
     }
