+++ probleme0xx/probleme053.cpp
@@ -16,8 +16,8 @@ ENREGISTRER_PROBLEME(53, "Combinatoric selections") {
     //
     // How many, not necessarily distinct, values of  nCr, for 1 ≤ n ≤ 100, are greater than one-million?
     size_t compteur = 0;
-    for (size_t n = 1; n <= 100; ++n) {
-        for (size_t p = 0; p <= n; ++p) {
+    for (unsigned long n = 1; n <= 100; ++n) {
+        for (unsigned long p = 0; p <= n; ++p) {
             if (mp_nombre::coefficient_binomial(n, p) > 1000000) ++compteur;
         }
     }
